+++ src/MySQL_Generic.h
@@ -12,7 +12,7 @@
   Built by Khoi Hoang https://github.com/khoih-prog/MySQL_MariaDB_Generic
   Licensed under MIT license
   
-  Version: 1.7.0
+  Version: 1.7.1
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -26,13 +26,17 @@
   1.6.0   K Hoang      10/03/2022 Fix memory leak bug. Optimize code
   1.6.1   K Hoang      12/03/2022 Fix memory management bug
   1.7.0   K Hoang      12/03/2022 Convert to `h-only` style library
+  1.7.1   K Hoang      10/04/2022 Use Ethernet_Generic library as default. Support SPI1/SPI2 for RP2040/ESP32
  **********************************************************************************************************************************/
+
+#pragma once
 
 #ifndef MYSQL_GENERIC_H
 #define MYSQL_GENERIC_H
 
-#if ( USE_ETHERNET || USE_ETHERNET_LARGE || USE_ETHERNET2 || USE_ETHERNET3 || USE_ETHERNET_ENC || USE_ETHERNET_ESP8266 || \
-      USE_UIP_ETHERNET || USE_NATIVE_ETHERNET || USE_QN_ETHERNET || USE_CUSTOM_ETHERNET || USE_ETHERNET_PORTENTA_H7 )
+#if ( USE_ETHERNET_GENERIC || USE_ETHERNET_ENC || USE_ETHERNET_ESP8266 || USE_UIP_ETHERNET || USE_ETHERNET_LAN8742A ||  \
+      USE_ETHERNET_LAN8720 || USE_NATIVE_ETHERNET || USE_QN_ETHERNET || USE_CUSTOM_ETHERNET || USE_ETHERNET_PORTENTA_H7 || \
+      USE_CUSTOM_ETHERNET )
 
   #include "MySQL_Generic_Ethernet.h"