+++ src/MySQL_Generic.hpp
@@ -12,7 +12,7 @@
   Built by Khoi Hoang https://github.com/khoih-prog/MySQL_MariaDB_Generic
   Licensed under MIT license
   
-  Version: 1.7.0
+  Version: 1.7.1
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -26,21 +26,24 @@
   1.6.0   K Hoang      10/03/2022 Fix memory leak bug. Optimize code
   1.6.1   K Hoang      12/03/2022 Fix memory management bug
   1.7.0   K Hoang      12/03/2022 Convert to `h-only` style library
+  1.7.1   K Hoang      10/04/2022 Use Ethernet_Generic library as default. Support SPI1/SPI2 for RP2040/ESP32
  **********************************************************************************************************************************/
+
+#pragma once
 
 #ifndef MYSQL_GENERIC_HPP
 #define MYSQL_GENERIC_HPP
 
-#define MYSQL_GENERIC_VERSION               		"1.7.0"
+#define MYSQL_GENERIC_VERSION               		"1.7.1"
 
 #ifndef MYSQL_MARIADB_GENERIC_VERSION
   #define MYSQL_MARIADB_GENERIC_VERSION       	("MySQL_MariaDB_Generic v" MYSQL_GENERIC_VERSION)
 
   #define MYSQL_MARIADB_GENERIC_VERSION_MAJOR    1
   #define MYSQL_MARIADB_GENERIC_VERSION_MINOR    7
-  #define MYSQL_MARIADB_GENERIC_VERSION_PATCH    0
+  #define MYSQL_MARIADB_GENERIC_VERSION_PATCH    1
 
-  #define MYSQL_MARIADB_GENERIC_VERSION_INT      1007000
+  #define MYSQL_MARIADB_GENERIC_VERSION_INT      1007001
 #endif
 
 #include <MySQL_Generic_Connection.h>