+++ src/MySQL_Generic_Encrypt_Sha1_Impl.h
@@ -17,7 +17,7 @@
   Built by Khoi Hoang https://github.com/khoih-prog/MySQL_MariaDB_Generic
   Licensed under MIT license
   
-  Version: 1.7.0
+  Version: 1.7.1
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -31,7 +31,10 @@
   1.6.0   K Hoang      10/03/2022 Fix memory leak bug. Optimize code
   1.6.1   K Hoang      12/03/2022 Fix memory management bug
   1.7.0   K Hoang      12/03/2022 Convert to `h-only` style library
+  1.7.1   K Hoang      10/04/2022 Use Ethernet_Generic library as default. Support SPI1/SPI2 for RP2040/ESP32
  **********************************************************************************************************************************/
+
+#pragma once
 
 #ifndef MYSQL_GENERIC_ENCRYPT_SHA1_IMPL_H
 #define MYSQL_GENERIC_ENCRYPT_SHA1_IMPL_H