+++ changelog.md
@@ -12,6 +12,7 @@
 ## Table of Contents
 
 * [Changelog](#changelog)
+	* [Release v1.7.1](#release-v171)
 	* [Release v1.7.0](#release-v170)
   * [Release v1.6.1](#release-v161)
   * [Release v1.6.0](#release-v160)
@@ -32,6 +33,15 @@
 ---
 
 ## Changelog
+
+### Releases v1.7.1
+
+1. Use new [Ethernet_Generic library](https://github.com/khoih-prog/Ethernet_Generic) as default for W5x00.
+2. Support SPI2 for ESP32
+3. Add support to SPI1 for RP2040 using [arduino-pico core](https://github.com/earlephilhower/arduino-pico)
+4. Use new function `waitForLink()` for Teensy 4.1 QNEthernet library v0.14.0+
+5. Rewrite all the examples to support those new features
+6. Update `Packages' Patches`
 
 ### Releases v1.7.0